+++ src/index.js
@@ -1,3 +1,10 @@
+// untouched material-ui
+export { default as FormControl } from '@material-ui/core/FormControl'
+export { default as FormControlLabel } from '@material-ui/core/FormControlLabel'
+export { default as FormGroup } from '@material-ui/core/FormGroup'
+export { default as FormLabel } from '@material-ui/core/FormLabel'
+export { default as RadioGroup } from '@material-ui/core/RadioGroup'
+
 // theme
 export { DefaultTheme, ThemeProvider } from './theme'
 
@@ -6,17 +13,16 @@ export { default as Alert } from './alert'
 export { default as Button } from './button'
 export { default as Card } from './card'
 export { default as Carousel } from './image-slider/carousel'
+export { default as Checkbox } from './checkbox'
 export { default as Chip } from './chip'
-export { default as DataTable } from './data-table' // deprecated by Table
 export { default as DynamicButton } from './dynamic-button'
 export { default as ImageSlider } from './image-slider'
 export { default as List } from './list'
 export { default as Loader } from './loader'
 export { default as QuickFilters } from './quick-filters'
+export { default as Radio } from './radio'
 export { default as StyledCardContainer } from './styled-card-container'
-export { default as StyledCheckbox } from './styled-checkbox'
-export { default as StyledRadio } from './styled-radio'
-export { default as StyledSwitch } from './styled-switch'
+export { default as Switch } from './switch'
 export { default as Table } from './table'
 export { default as TabPanels } from './tab-panels'
 export { default as TextField } from './text-field'
@@ -24,6 +30,10 @@ export { default as Typography } from './typography'
 export { default as WidgetStats } from './widget-stats'
 
 // deprecated
-// export { default as DynamicDataTable } from './dynamic-data-table' // deprecated by Table
-// export { default as WidgetNumber } from './widget-number' // deprecated by WidgetStats
-// export { default as WidgetTrend } from './widget-trend' // deprecated by WidgetStats
+export { default as DataTable } from './deprecated/data-table' // deprecated by Table
+export { default as DynamicDataTable } from './deprecated/dynamic-data-table' // deprecated by Table
+export { default as StyledCheckbox } from './deprecated/styled-checkbox' // deprecated by Checkbox
+export { default as StyledRadio } from './deprecated/styled-radio' // deprecated by Radio
+export { default as StyledSwitch } from './deprecated/styled-switch' // deprecated by Switch
+export { default as WidgetNumber } from './deprecated/widget-number' // deprecated by WidgetStats
+export { default as WidgetTrend } from './deprecated/widget-trend' // deprecated by WidgetStats