+++ stories/deprecated/styled-switch.stories.js
@@ -1,10 +1,10 @@
 import React from 'react'
 import { FormControlLabel } from '@material-ui/core'
 
-import { StyledSwitch } from '../src/index'
+import { StyledSwitch } from '../../src/index'
 
 export default {
-  title: 'Inputs/StyledSwitch',
+  title: 'Deprecated/StyledSwitch',
   component: StyledSwitch,
   args: {
     control: <StyledSwitch />,